task_1/ETL/requirements.txt
kafka-python==2.0.2
orjson==3.9.10
pydantic==2.5.3
pydantic-settings==2.1.0
pydantic_core==2.14.6
python-dotenv==1.0.0
clickhouse-driver==0.2.7
confluent-kafka==2.3.0